+++ app/javascript/app/controllers/order_controller.js
@@ -1,11 +1,19 @@
 import { Controller } from "stimulus"
 
 export default class extends Controller {
-  static targets = ["select"]
+  static targets = ["form", "select"]
 
-  connect(){}
-
-  initialize(){
+  connect() {
   }
 
+  initialize() {
+  }
+
+  reload(event) {
+    let form = this.formTarget;
+    form.submit();
+    document.getElementsByClassName("spinner-background")[0].style.display = 'flex';
+  }
 }
+
+